Fix Bring back twitter
* Fix the patch structure. (Make it a single resource patch.)
* Fix the issue that Japanese arrays.xml is not modified.
* Introduce region comments. This is often used in the official ReVanced project.

TODO: Logs are printed (not fixed in this commit)
